fix(burn/13): resolve merge conflict issues
- Add & ref to ndarray_device() call for Burn 0.21 TensorCreationOptions API
- Remove unused build_preflight import in diffusion.rs test module
- Add #[allow(dead_code)] to burn/10+11 scaffolding (load_diffusion_weights,
  load_tensor, load_tensor_opt, VaeWeightData, SdxlVaeDecoderWeights,
  VaeBlockWeights)
- Fix clippy: #[allow(clippy::too_many_arguments)] on sample_sdxl,
  remove redundant cfg as f32 cast, replace useless format! with to_string()
- Run cargo fmt

+++ crates/inference-backends/burn/src/operation/diffusion.rs
@@ -184,7 +184,7 @@ pub fn execute_diffusion_sample(
         &positive_payload,
         &negative_payload,
         steps,
-        cfg as f32,
+        cfg,
         seed,
         backend,
     )?;
@@ -286,7 +286,6 @@ mod tests {
     use super::*;
     use crate::config::BurnBackendConfig;
     use crate::models::stable_diffusion::sdxl::{BurnLoadedModelBundle, BurnLoadedSdxlBundle};
-    use crate::operation::text::build_preflight;
     use crate::profile::BACKEND_LABEL;
     use reimagine_core::model::{ModelId, NodeId, RunId, WorkflowId, WorkflowVersion};
     use reimagine_inference::{
